bit reworked container logwatch
